requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1

bleach~=6.0.0